requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
vllm==0.2.1
python-jose==3.3.0
python-multipart==0.0.6
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3
apscheduler==3.10.4
